+++ AltCover.PowerShell/Command.fs
@@ -138,14 +138,6 @@ type InvokeAltCoverCommand() =
   member val LcovReport = String.Empty with get, set
 
   /// <summary>
-  /// <para type="description">File path for JSON format version of the collected data</para>
-  /// </summary>
-  [<Parameter(ParameterSetName = "Runner", Mandatory = false, ValueFromPipeline = false,
-              ValueFromPipelineByPropertyName = false)>]
-  [<SuppressMessage("Microsoft.Naming", "CA1704", Justification = "Json is a name")>]
-  member val JsonReport = String.Empty with get, set
-
-  /// <summary>
   /// <para type="description">One or more of minimum acceptable statement (S), branch (B) or method (M)/alternative method (AM) coverage percentage (integer, 1 to 100) or maximum acceptable CRAP/alternative CRAP score (C/AC followed by integer, 1 to 255) e.g. M80AM70C40AC100B50. If the value starts with a number, a leading S is assumed. If any threshold is specified more than once, the last instance is assumed -- so 25S50 counts as S50. Zero/absent values are ignored. If a coverage result is below threshold, or the CRAP score is above threshold, the return code of the process is the largest abs(threshold - actual) rounded up to the nearest integer.</para>
   /// </summary>
   [<Parameter(ParameterSetName = "Runner", Mandatory = false, ValueFromPipeline = false,
@@ -545,8 +537,7 @@ type InvokeAltCoverCommand() =
         CommandLine = self.CommandLine
         ExposeReturnCode = not self.DropReturnCode.IsPresent
         SummaryFormat = String(formatString)
-        Verbosity = self.Verbosity
-        JsonReport = self.JsonReport }
+        Verbosity = self.Verbosity }
 
   member private self.Prepare() =
     let showStatic = [| "-"; "+"; "++ " |]