+++ ReleaseNotes.md
@@ -17,6 +17,7 @@ A. Start with the Quick Start guide : https://github.com/SteveGilham/altcover/wi
   * Because int64 doesn't fit as a double, tracking-related timestamps represented as Base64Encoded strings `Convert.ToBase64String(BitConverter.GetBytes(IPAddresss.HostToNetworkOrder(ticks)))`
   * Native JSON to LCov or Cobertura conversions are currently not supported
 * [Breaking] with the new JSON output the `-x, --xmlReport` argument or equivalent just becomes `-r, --report`
+* [Breaking] the stop-gap `--jsonReport` option from v7.6.any is withdrawn
 * [BUGFIX] Issue 122 -- rework the method name tokenization for extracting the `returnType (argumentList)` signature extraction in the Cobertura output, fixing an off-by-one error that generated `returnType argumentList)` without the `(` as well as the headline exception.
 
 # 7.6.812 (Genbu series release 15)